+++ include/onmt/Model.hxx
@@ -7,6 +7,7 @@ namespace onmt
 
   template <typename MatFwd, typename MatIn, typename MatEmb, typename ModelT>
   Model<MatFwd, MatIn, MatEmb, ModelT>::Model(const std::string& filename)
+    : _module_factory(_profiler)
   {
     THFile* tf = THDiskFile_new(filename.c_str(), "r", 0);
     THFile_binary(tf);
@@ -168,6 +169,12 @@ namespace onmt
       return default_value;
 
     return static_cast<T>(it->second);
+  }
+
+  template <typename MatFwd, typename MatIn, typename MatEmb, typename ModelT>
+  void Model<MatFwd, MatIn, MatEmb, ModelT>::enable_profiling()
+  {
+    _profiler.enable();
   }
 
 }