+++ include/onmt/nn/Module.h
@@ -4,6 +4,8 @@
 #include <string>
 #include <vector>
 
+#include "onmt/Profiler.h"
+
 namespace onmt
 {
   namespace nn
@@ -14,6 +16,7 @@ namespace onmt
     {
     public:
       Module(const std::string& name);
+      Module(const std::string& name, bool profile);
       virtual ~Module() {}
 
       std::vector<MatFwd> forward(std::vector<MatFwd>& input) const;
@@ -30,10 +33,13 @@ namespace onmt
       virtual std::string get_details() const;
 
       void set_custom_name(const std::string& custom_name);
+      void set_profiler(Profiler& profiler);
 
     protected:
       std::string _name;
       std::string _custom_name;
+      bool _profile;
+      mutable Profiler* _profiler;
 
       std::function<void(std::vector<MatFwd>&)> _post_process;
     };