+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+
+* Add module profiling
+
 ## [v0.4.1](https://github.com/OpenNMT/CTranslate/releases/tag/v0.4.1) (2017-03-08)
 
 ### Fixes and improvements